+++ scripts/deployment/index.js
@@ -1,2 +1,5 @@
 export { ActorFolderService } from "./folder-service.js";
 export { EncounterDeploymentService } from "./deployment-service.js";
+export { SceneDeploymentService } from "./scene-deployment-service.js";
+
+export { InteractiveTokenPlacementService } from "./interactive-token-placement-service.js";